Replace M_PI with a constexpr equivalent
When working with MSVC, _USE_MATH_DEFINES must be defined before the
first inclusion of <cmath> for M_PI to be available for use. This
requirement is problematic for projects that use <cmath> as the first
inclusion of <cmath> isn't always obvious or easy to track down and
defining <cmath> globally leads to redefinition warnings. This commit
addresses the issue by replacing all uses of M_PI with a constexpr
value for pi defined in the nanoflann namespace.