Install nanoflann architecture-independently
This PR installs the nanoflann.pc to /usr/share/pkgconfig, where it will
be found for all architectures, and adds the ARCH_INDEPENDENT option to
write_basic_package_version_file().

The latter requires at least CMake 3.14 to work. If this is not
acceptable for some reason, I can rewrite this with a known workaround
to set CMAKE_SIZE_OF_VOID_P to "" temporarily.