Use target compile features instead of CXX_STANDARD
This is a very nice feature of CMake that was added in version 3.1.

Instead of forcing C++11 with CXX_STANDARD, which might actually
*downgrade* the compiler from a newer standard, we just tell CMake which
C++ features we need, and CMake will automatically choose the proper
standard (or complain if the compiler does not support it).

Additionally, I added the feature list as a target property, so it will
be put in the installed CMake config script and propagate to library
users. This way, CMake will automatically ensure that any project that
includes nanoflann is compiled with at least C++11.